upload: rename app component
